// lib/generated-client/src/item/schemas/item/update-item/UpdateItemResponse.ts
import { z } from "zod";

export const UpdateItemResponseSchema = z.object({
	code: z.number().int().optional(),
	message: z.string().optional(),
	data: z.object({
	"@entity": z.literal("org.quicko.qpon.item").optional(),
	item_id: z.string().optional(),
	name: z.string().optional(),
	description: z.string().optional(),
	custom_fields: z.object({}).optional(),
	external_id: z.string().optional(),
	created_at: z.string().datetime({
	offset: true,
}).optional(),
	updated_at: z.string().datetime({
	offset: true,
}).optional(),
}).optional(),
});

export type UpdateItemResponse = z.infer<typeof UpdateItemResponseSchema>;